Send new location to server, started notification implementatie

+++ CeMeOiOS/CeMeOiOS/RequiredContactsNewMeetingViewController.m
@@ -100,7 +100,7 @@
  */
 - (IBAction)ContactSwitched:(id)sender {
     UISwitch *contactSwitch = (UISwitch *)sender;
-    int rowIndex = contactSwitch.tag;
+    NSInteger rowIndex = contactSwitch.tag;
     ContactCell *cell = (ContactCell *)[self.tableView cellForRowAtIndexPath:[NSIndexPath indexPathForRow:rowIndex inSection:0]];
     cell.contactSwitch.on = contactSwitch.on;
